projects/P02_deduplication_tmp/pyproject.toml
[tool.poetry]
name = "p02-deduplication"
version = "0.1.0"
description = ""
authors = ["Gaetan Pinon <g.pinon@quinten-health.com>"]
readme = "README.md"
packages = [
    { include = "p02_deduplication", from = "src" },
    { include = "scripts", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.11"
snakemake = "^8.29.2"
click = "^8.1.8"
registry-catalog-data-science = {path = "../..", develop = true}
ipykernel = "^6.29.5"
pandas = "^2.2.3"
tqdm = "^4.67.1"
openpyxl = "^3.1.5"
mistralai = "^1.5.0"
mistral-common = "^1.5.2"
python-dotenv = "^1.0.1"
dvc = "^3.59.0"
openai = "^1.61.1"
matplotlib = "^3.10.0"
weaviate-client = "^4.11.0"
dvc-s3 = "^3.2.0"
sentence-transformers = "^3.4.1"
scikit-learn = "^1.6.1"
streamlit = "^1.42.2"
plotly = "^6.0.0"
umap-learn = {version = "^0.5.7", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
